Remove clearing of summary_box_canvas.style.display
Removed because summary_box_canvas is part of the embedded NgChm and
we don't want to messing with the NgChm contents other than through an
NgChm API call.

I could not figure out what this was supposed to do.  Removing it seems
to have no effect.  The display of the summary_box_canvas has been
controlled by adding and removing the 'hide' class for a long time.