Add ?v=version parameter for opening HTML pages. Fixes issue #55.
I used a version number instead of an md5sum for simplicity. This will not
affect caching effectiveness for releases.  You may want to use a private
version during development and testing.

Version number references are inserted into both the javascript code
(NGCHM_GUI_Util.js) and the Java code (BuilderVersion.java) during build
by the versioner-v2.sh script.

Version numbers are added to most HTML page references by the gotoPage
function in the above Javascript file.  There are other two places HTML
pages are opened.

First is the index.html page.  This was previously handled by adding a
welcome-file entry to WEB-INF/web.xml, but this can only be to a static
file. This was replaced by a servlet that redirects to the first HTML page
with the version number parameter added.

Second is the Upload_Matrix.html page, but that is already opened with a
nonce parameter that also serves as a cache breaker.